Refactor: Flatten src/core/vector module hierarchy
- Merged `constants.rs`, `types.rs`, `validation.rs`, `metric.rs`, and `ops.rs` into `src/core/vector/mod.rs` to reduce file fragmentation and boilerplate.
- Updated imports in `src/core/vector/sparse.rs` to reference merged constants.
- Preserved `simd.rs` (internal) and `sparse.rs` (distinct logic) as separate files.
- Deleted the original fragmented files.
- Verified functionality via `cargo test core::vector`.

This reduction aligns with the 'Razor' persona directives to flatten unnecessary hierarchies and simplify code navigation.
