+++ .jules/razor.md
@@ -7,3 +7,8 @@
 **Bloat:** `ColdStorage` trait (Single-implementation abstraction used only by `RedbColdStorage`).
 **Cut:** Deleted the `ColdStorage` trait and `cold_storage.rs` module. Refactored all consumers to use the concrete `RedbColdStorage` struct directly.
 **Saved:** ~300 lines of boilerplate (trait definitions, mock implementations, duplicate imports) + removed dynamic dispatch overhead.
+
+## [Reduction]
+**Bloat:** Fragmented `core::vector` module (5 tiny files for simple utilities).
+**Cut:** Merged `constants.rs`, `types.rs`, `validation.rs`, `metric.rs`, and `ops.rs` into `mod.rs`.
+**Saved:** 5 files, ~50 lines of boilerplate (imports/re-exports), and simplified module navigation.
